+++ src/pages/image_drop/color_thief_panel.rs
@@ -1,22 +1,18 @@
-use adw::prelude::*;
-use adw::subclass::prelude::*;
+use adw::{prelude::*, subclass::prelude::*};
 use gtk::{glib, glib::clone, glib::Receiver, glib::Sender, CompositeTemplate, gio::ListStore};
 
-use std::cell::{RefCell, Cell};
-use std::thread;
-use log::debug;
+use std::{cell::{RefCell, Cell}, thread};
 use color_thief::{get_palette, ColorFormat};
-
-use crate::toasts::add_error_toast;
-use crate::util::copy_color;
-use crate::i18n::i18n;
+use log::debug;
 
 use crate::dialog::save_palette_dialog::SavePaletteDialog;
+use crate::toasts::add_error_toast;
+use crate::i18n::i18n;
+use crate::util::copy_color;
 
 use super::dropped_image::DroppedImage;
 use super::extracted_color::ExtractedColor;
 use super::extracted_color_row::ExtractedColorRow;
-
 
 #[derive(Clone, Debug)]
 pub enum ExtractionAction {
@@ -31,7 +27,6 @@ mod imp {
     #[derive(Debug, Default, CompositeTemplate)]
     #[template(resource = "/io/github/nate_xyz/Paleta/color_thief_panel.ui")]
     pub struct ColorThiefPanelPriv {
-        
         #[template_child(id = "image_bin")]
         pub image_bin: TemplateChild<adw::Bin>,
 
@@ -62,8 +57,6 @@ mod imp {
 
         pub sender: RefCell<Option<Sender<ExtractionAction>>>,
         pub receiver: RefCell<Option<Receiver<ExtractionAction>>>,
-
-
     }
 
     #[glib::object_subclass]
@@ -147,28 +140,31 @@ impl ColorThiefPanel {
             }
         }));
 
-        imp.save_button.connect_clicked(clone!(@strong self as this => @default-panic, move |_button| {
-            this.save_palette();
-        }));
+        imp.save_button.connect_clicked(
+            clone!(@strong self as this => @default-panic, move |_button| {
+                this.save_palette();
+            })
+        );
 
         imp.quality_dropdown.set_selected(1);
-
         imp.count_amount.set(imp.count_amount_spin.value());
         imp.quality.set(self.quality());
 
-        imp.count_amount_spin.connect_value_changed(clone!(@strong self as this => @default-panic, move |spin_button| {
-            this.imp().count_amount.set(spin_button.value());
-            this.start_extraction()
-        }));
+        imp.count_amount_spin.connect_value_changed(
+            clone!(@strong self as this => @default-panic, move |spin_button| {
+                this.imp().count_amount.set(spin_button.value());
+                this.start_extraction()
+            })
+        );
         
-        imp.quality_dropdown.connect_selected_notify(clone!(@strong self as this => @default-panic, move |_drop_down| {
-            this.imp().quality.set(this.quality());
-            this.start_extraction()
-        }));
-
+        imp.quality_dropdown.connect_selected_notify(
+            clone!(@strong self as this => @default-panic, move |_drop_down| {
+                this.imp().quality.set(this.quality());
+                this.start_extraction()
+            })
+        );
 
         self.setup_channel();
-    
     }
 
     fn setup_channel(&self) {
@@ -179,8 +175,6 @@ impl ColorThiefPanel {
             clone!(@strong self as this => move |action| this.clone().process_action(action)),
         );
     }
-
-
 
     fn quality(&self) -> u8 {
         match self.imp().quality_dropdown.selected() {
@@ -193,9 +187,9 @@ impl ColorThiefPanel {
 
     pub fn set_image(&self, image: DroppedImage) {
         let imp = self.imp();
-        self.set_visible(true);
         imp.image_bin.set_child(Some(&image));
         imp.image.replace(Some(image));
+        self.set_visible(true);
         self.list_store().remove_all();
         self.start_extraction();
     }
@@ -230,22 +224,19 @@ impl ColorThiefPanel {
     }
 
     fn extraction_done(&self, colors: Option<Vec<(u8, u8, u8)>>) {
-        match colors {
-            Some(colors) => {
-                let imp = self.imp();
-                imp.spinner.stop();
-                imp.spinner.set_visible(false);
-                imp.palette_box.set_visible(true);
-                imp.list_store.remove_all();
+        if let Some(colors) = colors {
+            let imp = self.imp();
+            imp.spinner.stop();
+            imp.spinner.set_visible(false);
+            imp.palette_box.set_visible(true);
+            imp.list_store.remove_all();
 
-                for rgba in colors {
-                    imp.list_store.append(&ExtractedColor::new(rgba))
-                }
-                    
-            },
-            None => {
-                add_error_toast(i18n("Unable to extract colors from image."));
-            },
+            for rgba in colors {
+                imp.list_store.append(&ExtractedColor::new(rgba))
+            }
+        } else  {
+            add_error_toast(i18n("Unable to extract colors from image."));
+
         }
         debug!("extraction_done");
     }
@@ -275,9 +266,6 @@ impl ColorThiefPanel {
     fn list_store(&self) -> &ListStore {
         &self.imp().list_store
     }
-
-
-
 }
 
 fn color_format(has_alpha: bool) -> ColorFormat {
@@ -310,5 +298,3 @@ pub fn load_palette_from_bytes(pixbuf_bytes: &[u8], alpha: ColorFormat, count: u
     }
     None
 }
-
-
